+++ src/utils/PreferencesManager.java
@@ -2,10 +2,6 @@ package utils;
 import static messages.Message.error;
 import static messages.Message.info;
 import static messages.Message.warning;
-import gui.ApplicationScreen;
-import gui.CustomColorChooser;
-import gui.View;
-import interfaces.UIPreferences;
 import java.awt.Color;
 import java.awt.Font;
 import java.io.BufferedWriter;
@@ -15,7 +11,6 @@ import java.io.FileOutputStream;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.ObjectOutputStream;
-import java.io.Serializable;
 import java.util.stream.Stream;
 import javax.swing.DefaultComboBoxModel;
 import javax.swing.JButton;
@@ -25,8 +20,14 @@ import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JSlider;
-import net.miginfocom.swing.MigLayout;
 import org.apache.commons.io.FileUtils;
+import extra.XString;
+import gui.ApplicationScreen;
+import gui.CustomColorChooser;
+import gui.View;
+import interfaces.UIPreferences;
+import net.miginfocom.swing.MigLayout;
+import serializable.Settings;
 @SuppressWarnings({"static-access", "serial"})
 public class PreferencesManager extends View implements UIPreferences{
 	private ApplicationScreen appFrame;
@@ -239,7 +240,7 @@ public class PreferencesManager extends View implements UIPreferences{
 			error(prefPanel, "There are no setiings saved by user");
 			return;
 		}
-		StringBuilder str = null;
+		XString str = null;
 		JFileChooser chooser = new JFileChooser();
 		chooser.setCurrentDirectory(new File("app"));
 		chooser.setDialogTitle("Select a folder to export preferences report");
@@ -262,7 +263,7 @@ public class PreferencesManager extends View implements UIPreferences{
 		try {
 			String ls ="\n";
 			BufferedWriter writer = new BufferedWriter(new FileWriter(f));		
-			str = new StringBuilder(); 
+			str = new XString(); 
 			str.append("------File Copy Manager Preferences------\n\n");
 			str.append("\tFont name:" + settings.getFontName()+ls+ "\tButton Font Size:"+ settings.getBtnSize()+ls);
 			str.append("\tLabel Font Size:"+ settings.getLblSize()+ls);
@@ -272,6 +273,12 @@ public class PreferencesManager extends View implements UIPreferences{
 			writer.close();
 		} catch (IOException exc) {
 			fh.log(exc);
+		}
+		finally {
+			if(f.exists())
+				info(prefPanel,"Preferences exported successfully");
+			else
+				error(prefPanel,"Preferences could not be exported");
 		}
 	}
 	public void deleteAppSettings() {
@@ -294,7 +301,7 @@ public class PreferencesManager extends View implements UIPreferences{
 			error( "Could not delete app settings");
 		}
 		if (!dir.exists())
-			info(prefPanel, "App settings deleted", "Success");
+			info(prefPanel, "App settings deleted");
 		else{
 			String msg = "App settings could not be deleted";
 			warning(msg);
@@ -320,66 +327,5 @@ public class PreferencesManager extends View implements UIPreferences{
 		loadPreferences();
 		updatePreview();
 		applySettings();	
-	}
-}
-@SuppressWarnings({"serial"})
-class Settings implements Serializable {
-	private Color bg, fg;
-	private String fontName;
-	private int lblSize = 12, btnSize = 18;
-	private Font[] fonts = UIPreferences.fonts;
-	public boolean isFontAvailable(){
-		for(Font f:fonts)
-			if(f.getFontName().equals(fontName))
-				return true;
-		return false;	
-	}
-	public int getFontIndex(String name){
-		for(int i=0,max=fonts.length;i<max;i++)
-			if(fonts[i].getFontName().equals(name))
-				return i;
-		return -1;	
-	}
-	public void setBgColor(Color e) {
-		bg = e;
-	}
-	public void setFgColor(Color e) {
-		fg = e;
-	}
-	public Color getBgColor() {
-		/**
-		 * Default background color for most swing components
-		 * */
-		return bg != null ? bg:new Color(238,238,238);
-	}
-	public Color getFgColor() {
-		/**
-		 * Default foreground color for most swing components
-		 * */
-		return fg != null ? fg:new Color(51,51,51);
-	}
-	public int getLblSize() {
-		return lblSize;
-	}
-	public void setLblSize(int lblSize) {
-		this.lblSize = lblSize;
-	}
-	public int getBtnSize() {
-		return btnSize;
-	}
-	public void setBtnSize(int btnSize) {
-		this.btnSize = btnSize;
-	}
-	public String getFontName() {
-		return fontName;
-	}
-	public void setFontName(String fontName) {
-		this.fontName = fontName;
-	}
-	public Font getButtonFont() {
-		return new Font(fontName, Font.PLAIN, this.btnSize);
-	}
-	public Font getLabelFont() {
-		return new Font(fontName, Font.PLAIN, lblSize);
 	}
 }